requirements.txt
supabase
openai
streamlit
ipykernel
langgraph
langchain
langchain-openai
langchain-ollama
python-dotenv
tqdm
plotly
